requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cryptography==42.0.4
debugpy==1.6.6
Flask==2.3.2
Flask-Login==0.6.3
Flask-Migrate==4.0.6
Flask-SQLAlchemy==3.1.1
packaging==23.0
python-dotenv==1.0.0
python-editor==1.0.4
python-json-logger==2.0.7
requests==2.31.0
requests-futures==1.0.0
requests-oauthlib==1.3.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
tink==1.9.0